Restore location directory to exact state from dade9e5236bcf9d95653c6c03965491bda065d68
