+++ location/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "orb-location"
+version = "0.0.0"
+publish = false
+
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+clap.workspace = true
+eyre.workspace = true
+color-eyre.workspace = true
+orb-telemetry.workspace = true
+reqwest = { workspace = true, features = ["blocking", "json"] }
+serde.workspace = true
+serde_json.workspace = true
+serialport = "4.6.1"
+thiserror = "1.0"
+tracing.workspace = true
+tracing-subscriber.workspace = true
+wpactrl = "0.5.1"
+orb-endpoints.workspace = true
+orb-info = { workspace = true, features = ["orb-id", "orb-token"] }
+orb-security-utils = { workspace = true, features = [ "reqwest", "blocking" ] }
+chrono = { version = "0.4", features = ["serde"] }
+zbus = { workspace = true }
+tokio = { workspace = true, features = ["rt-multi-thread", "macros", "sync"] }
+tokio-util = { workspace = true }
