--- location/cellcom/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "orb-cellcom"
-version = "0.0.0"
-authors = ["Galileo Daras <galileo@getcoffee.io>"] 
-description = "API for the Quectel EC25 module"
-publish = false
-
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-
-[features]
-serde = ["dep:serde"]
-
-[dependencies]
-eyre.workspace = true
-serde = { workspace = true, features = ["derive"], optional = true } 
-serialport.workspace = true
-tracing.workspace = true
-
-[dev-dependencies]
-clap = { workspace = true, features = ["derive"] }
