--- location/cellcom/README.md
@@ -1,7 +1,0 @@
-# orb-cellcom
-
-This code provides a wrapper for the Quectel EC25 module.
-
-Important docs:
-- https://files.pine64.org/doc/datasheet/project_anakin/LTE_module/Quectel_EC25&EC21_QuecCell_AT_Commands_Manual_V1.1.pdf
-- https://forums.quectel.com/t/neighbor-cell-cell-id/36317
