--- location/daemon/Cargo.toml
@@ -1,39 +1,0 @@
-[package]
-name = "orb-location"
-version = "0.0.0"
-description = """
-Gathers data from wpa_supplicant and the Quectel EC25 module to resolve an accurate
-position estimate for the device.
-"""
-authors = [
-	"Michael Preibisch <Michael-Preibisch@users.noreply.github.com>",
-	"Galileo Daras <galileo@getcoffee.io>",
-]
-publish = false
-
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-
-[dependencies]
-chrono.workspace = true
-clap.workspace = true
-color-eyre.workspace = true
-eyre.workspace = true
-orb-cellcom = { workspace = true, features = ["serde"] }
-orb-endpoints.workspace = true
-orb-google-geolocation-api.workspace = true
-orb-info = { workspace = true, features = ["orb-id"] }
-orb-location-wifi.workspace = true
-orb-security-utils = { workspace = true, features = [ "reqwest", "blocking" ] }
-orb-telemetry = { workspace = true, features = [] }
-reqwest = { workspace = true, features = ["blocking", "json"] }
-serde.workspace = true
-serde_json.workspace = true
-thiserror.workspace = true
-tokio-util.workspace = true
-tokio.workspace = true
-tracing-subscriber.workspace = true
-tracing.workspace = true
-zbus.workspace = true
