--- location/daemon/README.md
@@ -1,5 +1,0 @@
-# Cellcom
-
-A daemon that gathers data using `orb-location-wpa-supplicant` and `orb-cellcom`, and
-reports it our the backend.
-
