--- location/google-geolocation-api/README.md
@@ -1,5 +1,0 @@
-# orb-google-geolocation-api
-
-This crate exposes types to interact with [Google's Geolocation Api][api].
-
-[api]: https://developers.google.com/maps/documentation/geolocation/requests-geolocation
