+++ location/src/cell/parser.rs
@@ -1,5 +1,5 @@
-use crate::data::{NeighborCell, ServingCell};
-use crate::types::{
+use crate::cell::data::{NeighborCell, ServingCell};
+use crate::cell::types::{
     parse_ec25_serving_cell, parse_opt_i32, parse_opt_u32, Ec25ServingCell,
 };
 use eyre::{eyre, Result};
