+++ location/src/data.rs
@@ -1,16 +1,14 @@
-//! This module contains types that are not part of google's api but are used in builders for the
-//! google types.
-
-use orb_cellcom::{NeighborCell, ServingCell};
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Clone)]
+use crate::cell::{NeighborCell, ServingCell};
+
+#[derive(Debug, Serialize, Deserialize)]
 pub struct NetworkInfo {
     pub wifi: Vec<WifiNetwork>,
     pub cellular: CellularInfo,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct WifiNetwork {
     pub bssid: String,
     pub frequency: u32,
@@ -19,7 +17,7 @@ pub struct WifiNetwork {
     pub ssid: String,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct CellularInfo {
     pub serving_cell: ServingCell,
     pub neighbor_cells: Vec<NeighborCell>,
