+++ location/src/lib.rs
@@ -1,5 +1,10 @@
 pub mod backend;
+pub mod cell;
 pub mod config;
+pub mod data;
 pub mod errors;
 pub mod network_manager;
 pub mod service;
+pub mod wifi;
+
+pub use errors::{LocationError, Result};
