--- location/wifi/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "orb-location-wifi"
-version = "0.0.0"
-authors = [
-	"Galileo Daras <galileo@getcoffee.io>",
-	"Michael Preibisch <Michael-Preibisch@users.noreply.github.com>",
-]
-description = "(Another) wpa-supplicant interface"
-publish = false
-
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-
-[dependencies]
-eyre.workspace = true
-orb-google-geolocation-api.workspace = true
-serde = { workspace = true, features = ["derive"] }
-serde_json.workspace = true
-tracing.workspace = true
-wpactrl = "0.5.1"
